requirements.txt
Django==4.2.4
djangorestframework==3.14.0
openai==0.27.8
pip==23.0.1
python-dotenv==1.0.0
